Database and Fully functional CRUD operations for resources, users, and comments
This version of the backend implements the GET and POST CRUD operations for the different entities that our project will be handling. Our frontend will be able to fetch the existing resources, users, and comments from our database and display them. Our frontend will also be able to create resources, users, and comments using the corresponding attribute fields for each entity. This version of the backend also includes and new directory called Database that stores our tables and records, this is done in order for the data to persist after program termination.